Moved buildWatcher to a driver
