+++ adapters/http/box.go
@@ -7,21 +7,13 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/republicprotocol/atom-go/adapters/atoms"
-
 	"github.com/btcsuite/btcutil"
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	ethCrypto "github.com/ethereum/go-ethereum/crypto"
 
 	btcClient "github.com/republicprotocol/atom-go/adapters/clients/btc"
 	ethClient "github.com/republicprotocol/atom-go/adapters/clients/eth"
 	"github.com/republicprotocol/atom-go/adapters/config"
-	ax "github.com/republicprotocol/atom-go/adapters/info/eth"
-	net "github.com/republicprotocol/atom-go/adapters/networks/eth"
-	"github.com/republicprotocol/atom-go/adapters/store/leveldb"
-	wal "github.com/republicprotocol/atom-go/adapters/wallet/eth"
-	"github.com/republicprotocol/atom-go/services/store"
 	"github.com/republicprotocol/atom-go/services/swap"
 	"github.com/republicprotocol/atom-go/services/watch"
 	"github.com/republicprotocol/atom-go/utils"
@@ -36,25 +28,17 @@ type boxHttpAdapter struct {
 	watch  watch.Watch
 }
 
-func NewBoxHttpAdapter(config config.Config, keystr swap.Keystore) (BoxHttpAdapter, error) {
-	watcher, err := buildWatcher(config, keystr)
-	if err != nil {
-		return nil, err
-	}
-
+func NewBoxHttpAdapter(config config.Config, keystr swap.Keystore, watcher watch.Watch) BoxHttpAdapter {
 	return &boxHttpAdapter{
 		config: config,
 		keystr: keystr,
 		watch:  watcher,
-	}, nil
+	}
 }
 
 func (adapter *boxHttpAdapter) WhoAmI(challenge string) (WhoAmI, error) {
-
 	version := adapter.config.GetVersion()
-
 	suppCurrencies := adapter.config.GetSupportedCurrencies()
-
 	authorizedAddresses := adapter.config.AuthorizedAddresses
 
 	boxInfo := BoxInfo{
@@ -141,69 +125,6 @@ func (adapter *boxHttpAdapter) PostOrder(order PostOrder) (PostOrder, error) {
 		order.OrderID,
 		MarshalSignature(sig65),
 	}, nil
-}
-
-func buildWatcher(config config.Config, kstr swap.Keystore) (watch.Watch, error) {
-	ethConn, err := ethClient.Connect(config)
-	if err != nil {
-		return nil, err
-	}
-
-	btcConn, err := btcClient.Connect(config)
-	if err != nil {
-		return nil, err
-	}
-
-	keys, err := kstr.LoadKeys()
-	if err != nil {
-		return nil, err
-	}
-
-	ethKey := keys[0]
-	btcKey := keys[1]
-
-	_WIF, err := btcKey.GetKeyString()
-	if err != nil {
-		return nil, err
-	}
-
-	WIF, err := btcutil.DecodeWIF(_WIF)
-	if err != nil {
-		return nil, err
-	}
-
-	err = btcConn.Client.ImportPrivKey(WIF)
-	if err != nil {
-		return nil, err
-	}
-
-	owner := bind.NewKeyedTransactor(ethKey.GetKey())
-	owner.GasLimit = 3000000
-
-	ethNet, err := net.NewEthereumNetwork(ethConn, owner)
-	if err != nil {
-		return nil, err
-	}
-
-	ethInfo, err := ax.NewEthereumAtomInfo(ethConn, owner)
-	if err != nil {
-		return nil, err
-	}
-
-	ethWallet, err := wal.NewEthereumWallet(ethConn, *owner)
-	if err != nil {
-		return nil, err
-	}
-
-	atomBuilder := atoms.NewAtomBuilder(config, keys)
-
-	db, err := leveldb.NewLDBStore(config.StoreLocation())
-	if err != nil {
-		return nil, err
-	}
-	str := store.NewSwapState(db)
-	watcher := watch.NewWatch(ethNet, ethInfo, ethWallet, atomBuilder, str)
-	return watcher, nil
 }
 
 func (adapter *boxHttpAdapter) GetStatus(orderID string) (Status, error) {
